requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
easyocr==1.7.2
filelock==3.13.1
Flask==3.1.1
flask-cors==6.0.0
fsspec==2024.6.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.0.0
pyclipper==1.3.0.post6
pytesseract==0.3.13
python-bidi==0.6.6
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
setuptools==70.2.0
shapely==2.1.1
sympy==1.13.1
tifffile==2025.6.1
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
typing_extensions==4.12.2
urllib3==2.4.0
Werkzeug==3.1.3
